requirements.txt
python-telegram-bot==20.8
python-dotenv==1.0.1
requests==2.31.0
asyncio==3.4.3
